Move OL related measure logic to own controller
This introduces an own controller class offering the logic for the
measure tool, which is related to OpenLayers.
