+++ src/components/measuretool/MeasureWin.vue
@@ -29,18 +29,11 @@
 </template>
 
 <script>
-  import DrawInteraction from 'ol/interaction/Draw';
-  import {unByKey} from 'ol/Observable.js';
-  import VectorSource from 'ol/source/Vector';
-  import VectorLayer from 'ol/layer/Vector';
-  import Style from 'ol/style/Style';
-  import Stroke from 'ol/style/Stroke';
-  import Circle from 'ol/style/Circle';
-  import Fill from 'ol/style/Fill';
   import { DraggableWin } from '../../directives/DraggableWin';
   import { Mapable } from '../../mixins/Mapable';
   import MeasureTypeChooser from './MeasureTypeChooser';
   import MeasureResult from './MeasureResult';
+  import OlMeasureController from './OlMeasureController';
 
   export default {
     name: 'wgu-measuretool-win',
@@ -71,13 +64,16 @@
       show () {
         var me = this;
         if (me.show === true) {
-          me.addInteraction();
+          me.olMapCtrl.addInteraction(me.measureType, me.onMeasureVertexSet);
         } else {
-          me.removeInteraction();
+          me.olMapCtrl.removeInteraction();
         }
       },
       measureType () {
-        this.addInteraction();
+        var me = this;
+        // reset old geom
+        me.measureGeom = {};
+        me.olMapCtrl.addInteraction(me.measureType, me.onMeasureVertexSet);
       }
     },
     methods: {
@@ -95,112 +91,23 @@
        * This function is executed, after the map is bound (see mixins/Mapable)
        */
       onMapBound () {
-        this.createMeasureLayer();
-      },
-      /**
-       * Creates a vector layer for the measurement results and adds it to the
-       * map.
-       */
-      createMeasureLayer () {
         const me = this;
         const measureConf = me.$appConfig.modules[me.moduleName] || {};
-        // create a vector layer to
-        var source = new VectorSource();
-        var vector = new VectorLayer({
-          name: 'Measure Layer',
-          displayInLayerList: false,
-          source: source,
-          style: new Style({
-            fill: new Fill({
-              color: measureConf.fillColor || 'rgba(255, 255, 255, 0.2)'
-            }),
-            stroke: new Stroke({
-              color: measureConf.strokeColor || 'rgba(0, 0, 0, 0.5)',
-              width: 2
-            })
-          })
-        });
+        this.olMapCtrl = new OlMeasureController(me.map, measureConf);
 
-        me.map.addLayer(vector);
-
-        // make vector source available as member
-        me.source = source;
+        me.olMapCtrl.createMeasureLayer();
       },
       /**
-       * Creates and adds the necessary draw interaction and adds it to the map.
+       * Callback function executed when user sets a measure point on the map.
+       *
+       * @param  {ol/geom/Geometry} geom The geometry object of the map
        */
-      addInteraction (newMeasureType) {
-        const me = this;
-        const measureConf = me.$appConfig.modules[me.moduleName] || {};
-        // cleanup possible old draw interaction
-        if (me.draw) {
-          me.removeInteraction();
-        }
-
-        var type = (me.measureType === 'area' ? 'Polygon' : 'LineString');
-        var draw = new DrawInteraction({
-          source: me.source,
-          type: type,
-          style: new Style({
-            fill: new Fill({
-              color: measureConf.sketchFillColor || 'rgba(255, 255, 255, 0.2)'
-            }),
-            stroke: new Stroke({
-              color: measureConf.sketchStrokeColor || 'rgba(0, 0, 0, 0.5)',
-              lineDash: [10, 10],
-              width: 2
-            }),
-            image: new Circle({
-              radius: 5,
-              stroke: new Stroke({
-                color: measureConf.sketchVertexStrokeColor || 'rgba(0, 0, 0, 0.7)'
-              }),
-              fill: new Fill({
-                color: measureConf.sketchVertexFillColor || 'rgba(255, 255, 255, 0.2)'
-              })
-            })
-          })
-        });
-        me.map.addInteraction(draw);
-
-        var listener;
-        var sketch;
-        draw.on('drawstart', (evt) => {
-          // clear old measure features
-          me.source.clear();
-          // preserve sketch
-          sketch = evt.feature;
-
-          listener = me.map.on('click', (evt) => {
-            const geom = sketch.getGeometry();
-            // wrap geom into object, otherwise the injection into childs does
-            // not work. Maybe the OL object does not feel changed for Vue
-            me.measureGeom = {
-              geom: geom
-            };
-          });
-        }, me);
-
-        draw.on('drawend', () => {
-          // unset sketch
-          sketch = null;
-          unByKey(listener);
-        }, me);
-
-        // make draw interaction available as member
-        me.draw = draw;
-      },
-      /**
-       * Removes the current interaction and clears the values.
-       */
-      removeInteraction () {
-        var me = this;
-        if (me.draw) {
-          me.map.removeInteraction(me.draw);
-        }
-        if (me.source) {
-          me.source.clear();
-        }
+      onMeasureVertexSet (geom) {
+        // wrap geom into object, otherwise the injection into childs does
+        // not work. Maybe the OL object does not feel changed for Vue
+        this.measureGeom = {
+          geom: geom
+        };
       }
     }
   }
